+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Admin Default layout & Login
 - Customer Detail Page with Sign-Out feaature
 - Schema Admin (res loginAdmin,def type Admin,input AdminLoginInput)[@john]
+- Admin validator
 ### Removed 
 - Join now button
 ### Fixed